record & export adding default folder in selenium IDE
